requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
python-docx==1.1.2
PyPDF2==3.0.1
chromadb==0.5.5
sentence-transformers==3.1.1